+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -300,6 +300,8 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             // Dialog settings.
             dialogBackgroundSchemeColor: controller.dialogBackgroundSchemeColor,
             dialogRadius: controller.dialogBorderRadius,
+            useInputDecoratorThemeInDialogs:
+                controller.useInputDecoratorThemeInDialogs,
             timePickerDialogRadius: controller.dialogBorderRadius,
             timePickerElementRadius: controller.timePickerElementRadius,
             dialogElevation: controller.dialogElevation,